controllers,view,forms,dao,validator: implementing validation on pages.
Adding new classes, which are represented forms on the page. It was made
for easily editing the forms without any changes in the model.
Implementing various validation for each form.
Implementing validator for Person entity, which checks, if the person
with the same full name is existed.
Adding method for read the person by it full name.